+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Fixed
+- Fix dev example block info element to use "label" rather than "text".
+
 ## [1.0.0] - 2021-06-25
 
 ### Added